fix: compute_intrinsic_width/compute_intrinsic_height must return integer value

+++ src/pictex/nodes/image_node.py
@@ -6,7 +6,6 @@ which is essential for stretchable layout to properly size images.
 import skia
 from .container_node import ContainerNode
 from ..models import Style
-from ..utils import clone_skia_rect
 
 
 class ImageNode(ContainerNode):
@@ -15,19 +14,19 @@ class ImageNode(ContainerNode):
     def __init__(self, style: Style) -> None:
         super().__init__(style, [])
 
-    def compute_intrinsic_width(self) -> float:
+    def compute_intrinsic_width(self) -> int:
         """Get intrinsic width from background image."""
         image = self._get_background_image()
         if image:
-            return float(image.width())
-        return 0.0
+            return image.width()
+        return 0
     
-    def compute_intrinsic_height(self) -> float:
+    def compute_intrinsic_height(self) -> int:
         """Get intrinsic height from background image."""
         image = self._get_background_image()
         if image:
-            return float(image.height())
-        return 0.0
+            return image.height()
+        return 0
     
     def _get_background_image(self) -> 'skia.Image | None':
         """Get the skia image from background_image style."""
